% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_summary.R
\name{get_player_summary}
\alias{get_player_summary}
\title{get_player_summary}
\usage{
get_player_summary(team_detail_file)
}
\arguments{
\item{team_detail_file}{\itemize{
\item the full path to an html page with player match
level details for a single team
}}
}
\value{
\itemize{
\item a data frame of player summary data for the team described in \code{team_detail_file}
}
}
\description{
get_player_summary
}
\examples{
\dontrun{
team_detail_file <- here::here(
  "input",
  "GOWTL score details",
  "3.0 Division - 300 Club Everlasting Lobstoppers.html"
  )
get_player_summary(team_detail_file)
}
}
